Add files via upload
The goal of this project is to use the One-versus-Rest technique to convert multi class classification issues into multiple binary classification problems. The Seeds dataset is used in this task. Training and test splits are provided in CSV file format, and the performance of Perceptron and SVM models are compared. Provide accuracies and confusion matrix for both models, as well as creating OvR-Perceptron and OvR-SVM and testing them on the Seeds testing dataset, and using argmax aggregation technique to aggregate confidence scores and obtain the final label and obtain the performance